Run scheduled build at 8am Thursdays
Originally was set to run 24hrs later to avoid people just checking the
website but that seems unlikely and it is actually better to be able to
see the latest email right away e.g. if you see traffic coming from
Console and want to find out why. Not helpful if the "latest" email is
out of date! We always send Thursdays 8am so this will happen at the
same time.